backend/requirements.txt
# django and related
django~=5.1.1
django-allauth[socialaccount]~=64.2.0
django-environ~=0.11.2
django-unfold~=0.37.0

# job scheduler
django-q2~=1.6.2
blessed~=1.20.0  # optional requirement for django-q2
psutil~=6.0.0  # optional requirement for django-q2
croniter~=3.0.3  # optional requirement for django-q2

# file handling
django-storages[s3]~=1.14.4
whitenoise~=6.7.0

# emails
django-ses~=4.1.1

# misc
requests~=2.32.3
python-fossil-delta==18.11.27

# Temporary dependency until we finish the `ngohub` package
pycognito~=2024.5.1

# database
psycopg2-binary~=2.9.9

# prod packages
gunicorn~=23.0.0
gevent~=24.2.1
sentry-sdk[django]~=2.13.0
